examples/bookstore/requirements.txt
django>=4.2
factory-boy
faker
django-graph-walker
